began work on uni cog
